# feat: cache parsed/loaded APIs for faster access