This PR includes a number of improvements related to caching of APIs that **significantly** improve the startup speed of Restish. These changes include:

- Adding a new binary cache of parsed API & operation info with a 24 hour default expiry
- Move the Markdown syntax highlighting from the API loader into the help display template. For a large API like Github's, rendering several hundred Markdown descriptions was extremely slow and 99.9% of the time they were never used. Now it gets rendered if/when needed via `--help`.

In practice using the Github v3 OpenAPI we see a massive speed up (like 5% of the original time taken!):

```sh
# Previous version
$ time restish-old github --help
restish-old github --help  1.85s user 0.10s system 135% cpu 1.439 total

# New version
$ time restish github --help
restish github --help  0.05s user 0.02s system 84% cpu 0.082 total
```

This has serious implications for the usability of dynamic shell completions where you expect the tool to respond almost instantly!